thegraph: Attach the current schema to incoming queries
This way we always know which schema a query was made against. I'm still
debating whether this is good design. The problem is that multiple components
receive the combined schema from the schema provider at potentially different
times (right now: GraphQlServer and Store), so it is not guaranteed that the
GraphQL server and store will have received and reacted to the latest schema
update in sync. We'll see if we can come up with a good solution for this
(it could require to send an Arc<Schema> around instead of a cloned Schema).

Ultimately, the schema known by the server is what the client considers the
source of truth, so it does make a little more sense to associate the schema
of the server with the query, as opposed to the schema that the store knows
(or the query runner, perhaps).