thegraph-core: Implement basic query execution in the query runner
The query execution algorithm follows the GraphQL spec very closely. For more
information, see http://facebook.github.io/graphql/October2016/#sec-Execution.

This comes with a bunch of query and schema AST helpers and a flexible resolver
concept that gives us a lot of freedom in resolving entities, enum values,
scalar values and abstract types such as interfaces.

It won't be hard to make the query execution asynchronous based on the futures
library, but we'll leave that for later.