thegraph: Change Store trait to have a channel based API
For example, instead of calling

    let result = store.find(query);

callers now do this:

    let (result_sender, result_receiver) = oneshot::channel();

    let sink = store.request_sink();
    sink.send(StoreRequest::Find(query, result_sender)).wait().unwrap();

    receiver
      .and_then(|result| ...)

While this looks less convenient, it allows us to move all questions
of ownership and synchronization between find/get and set/delete into
the store. Components that want to access the store no longer need to
have a reference to the store directly. They can instead just obtain
and clone the request sink.