thegraph-runtime: Gracefully handle Ethereum call errors
We will at some point have to decide how _exactly_ we want to handle
errors that cause data source event handlers to fail (because such
failures could lead to inconsistent / invalid data), whether it is
retrying, rolling back the database and terminating the node or whatever.

For now, it seems ok to just carry on and log a big fat warning.

The Trap returned from the `ethereum_call` host function allows WASM
to return from the event handler early and pass the trap to the
caller, where we are now logging it.