graph, node, json-rpc: Add json-rpc admin server
Running by default on port 8020.
For now it only supports the `subgraph_add` operation.

Make a subgraph hash optional in the CLI,
when present add it through the json-rpc API.
A subgraph name may optionally be provided in
the format `name:hash`.

No authorization for now, the server is openly
exposed to requests.

The easiest way to communicate with the server
is through the REST API, with a request such as:

```
POST http://localhost:8020/subgraph_add/name/hash
Content-type: application/json
```

Hosting multiple subgraphs is not yet supported.