graph: Bump and re-export web3, drop ethereum-types dependency
ethereum-types, ethabi and web3 all export the same fundamental
Ethereum types (e.g. Address and H256). This commit re-exports
web3 (and bumps it at the same time), so we can just depend on
the graph crate in all other crates and use graph::web3::types::*
from there.

As a consequence, we can drop the dependency of ethereum-types
and limit our use of ethabi to a minimum, making it easier to
bump our dependencies without introducing versioning conflicts.