Use BlockTicks in GraphQL subscriptions
We used to listen to EntityChanges, but several of those can be fired while
processing a single block. It is enough to update a GraphQL subscription at
most once per block. To that end, we now use a BlockTick subscription which
ensures we do not update the subscription more than once a block.