Change how we generate events for entity changes
Previously, we sent an event to subscribers for every change that was made
to an entity in the database. To reduce the number of events the database
generates, we now batch changes into events, and generate only one such
event per block for block-based data. Non-block-based data, in particular
changes to the subgraph of subgraphs, produce an event for each DB
transaction in which subgraph metadata was changed.