Simplify enum EntityChangeOperation
We do not distinguish anymore between 'Added' and 'Updated', but collapse
those two into a 'Set' variant. The distinction between the two was
meaningless in the code, and with the way we now generate store events, we
can no longer say whether an operation results in an insert or update
without consulting the database.