store/postgres (StoreEvent): store changes as a HashSet
Using a HashSet instead of a Vec here more closely expresses what we are
trying to say: these are the entities that have been changed, without
implying any order. It also helps with the size of messages for blocks that
see a lot of changes to the same entities as they won't be added to the
event multiple times.