store/postgres: delete unneeded entity history
Get rid of the following in entity_history:

- all history for the 'subgraphs' subgraph
- all history where data_before = data_after
- the data_after column in entity_history
- any event_meta_data event that is not used by an entry in entity_history

The migration is a little more complicated than one might think because we
are trying to do this as fast as possible: instead of deleting from
entity_history and event_meta_data, we copy the 'good' data to a new
table. We also defer creating constraints and indexes until after the data
has been copied since bulk creation is faster than doing it on every row.

Fixes https://github.com/graphprotocol/graph-node/issues/807