core, datasource/etheruem, graph: Simplify log filter management
The subgraph's current log filter is now exclusively stored in the
`SubgraphState`. It is extended whenever new data sources have been
added. It is never passed or stored in `SubgraphInstance` or
`BlockStreamBuilder`. It is only passed to the `BlockStreamBuilder`
when it is time to build a new stream with a log filter.