store/postgres: Add `add_entity_history_record` helper
This takes a history event, a reversion flag, an entity key to create
the history record for and an operation type (e.g. `OperationType::Insert`).
It then creates an entry in the subgraph's `entity_history` table.

Special casing is applied for the "public.entities" and "subgraphs.entities"
tables, as their history records include the subgraph ID in the "subgraph"
column.