store/postgres: make sure only one graph-node runs database migrations
We had the problem that if multiple graph-nodes started up simultaneously,
all of them would try to run database migrations. Only one of them would
succeed, and the others would have their migrations rolled back. That
causes issues with slow or expensive migrations because it forces the
database to do a lot of unnecessary work.

With this change, only one graph-node will migrate the database, all the
others will wait for it to finish by blocking on a new
__graph_node_global_lock table.