store/postgres: Make subgraph creation not block writers
Creating a subgraph can block all other subgraph writers if the creation
itself has to wait for its lock, e.g., because of an ongoing autovacuum of
event_meta_data. That wait can be long (an hour or more in the case of
autovacuum)

We now do subgraph creation in such a way that we wait for a lock only a
short time (2s) and then sleep for a bit so that other write activity can
proceed.