graphql, graph: Make max value for `first` configurable (#1044)
This adds a `GRAPHQL_MAX_FIRST` environment that allows to set the
maximum value that is allowed for `first`. The default `first` value
remains at `100` to not accidentally push existing queries past the
complexity limit.

The error message for when `first` is too high now includes the
configured limit, e.g.

> Value of "first" must be between 1 and 1000