graphql: Guarantee at least one subscription result
The GraphQL over WebSocket spec demands that for every subscription that
is created, at least one result is sent back. To satisfy this, we're
injecting an empty store event in the source stream created when a
subscription is set up. This then triggers the subscription's query to
be executed once before waiting for _real_ store events.