store/postgres: Make sure missing notification does not panic listener
When the payload for a large notification was cleaned up too early and went
missing, the notification listener would panic because we
postgres::row::Rows::get panics when we try to access the nonexistent row
in the result.

We now check that the result is not empty before accessing values to avoid
that panic.

This partly addresses https://github.com/graphprotocol/graph-node/issues/1043