Use EntityCache in HostExports
Instead of using teh EntityCache only within the store to deduplicate
entity operations, and weed out operations with no effect, use it in the
runtime already. That has the advantage that we will never get the same
entity more than once from the store. Before, a `store.get` followed by a
`store.set` would have caused two lookups of the entity, one in the
runtime, and one in the store.