store: Reduce the size of string prefixes we index
The previous size of 2k was the largest we could store in a Postgres BTree
index, but it's excessive for supporting normal string operations; anything
beyond a few hundred bytes is text, and very unlikely to be accessed
through 'hard' string comparisons; it's much more likely that such text
will be accessed through word searches and similar for which the BTree
index is of no use anyway.

We lower the prefix size to 256 to reduce the size of the index for
attributes that do store large text entries.