ethereum: Handle Geth reverts and empty responses
We previously held back from accepting a `0x` from
Geth as a revert, because it was used for timeouts.

Geth has since fixed that an timeouts have a different
response, so `0x` is a determinstic response. It can
also be that the contract returned an empty response,
but that's also an error, even if not techinically a revert,
so we treat it the same.