store: Remove `guard` from `MetadataOperation::Update`
It was only used in `update_ethereum_block_pointer_operations`,
which in turn is only used in reverts or transact block.

For transact block, the guard was made totally redundant
as we now fetch the block in the same transaction.

For reverts, I was more conservative and simply added
an assert to replace the guard.