ethereum: Dynamic block range based on triggers per range
Use the amount of triggers found in the previous range
to adjust the next range, trying to meet a configured
target triggers per range. For dense subgraphs this
allows the complexity of ethereum calls and
memory usage to be kept under control, without
compromising speed for cheap subgraphs or
initial speed for subgraphs that are denser
towards the end of the chain.