chain, core, graph, mock: Avoid `Arc<Self>` in `EthereumAdapter`
This is done to allow the `EthereumAdapter` trait to be mocked easier
with mockall (see the next commit). mockall doesn't support `Arc<Self>`
right now.

The main change required to achieve this is to move the
`block_with_triggers` and `triggers_in_block` methods out of the trait
and move some parsing helpers into the `graph` crate.