chain/ethereum: Rewrite reorg logic
This rewrites the reorg logic to collect old and new blocks for a reorg
block by block, traversing parents, instead of collecting them
separately. The previous way of identifying the common ancestor of a
reorg and the local chain would consider ommers as being part of the
local chain, which they are not.