chain/ethereum: Drastically simplify reorg handling
Instead of collecting all old and new blocks to find the common
ancestor and revert old blocks, we simply revert the local head block
one block at a time and re-evaluate the situation (by polling the chain
head block again and deciding which blocks to look up next). Eventually,
this procedure will revert the local head back to the common ancestor.

For deep reorgs, this will be slow, but about 99% of Ethereum reorgs
have a depth of one, so this is something we can live with easily.