core: Fix retrying in LinkResolver::cat
Adding support for multiple IPFS clients added a non-retrying
`object.stat` into the mix, which unfortunately breaks the retry feature
of `LinkResolver`. We added this retry feature to work around
intermittent IPFS issues when starting subgraphs (e.g. retry when
hitting an IPFS rate limit or temporary DNS error), reducing the chances
of subgraphs failing to start at all.

This fixes the regression by wrapping the logic that identifies the best
IPFS client for an IPFS file in a retry future.