core: Retry the entire LinkResolver::cat body
This makes the retry logic easier to follow and means we are retrying
whenever one of the two steps (identify the best IPFS client, retrieve
the file) fails. Otherwise we might run into situations where we find an
IPFS client but by the time we try to retrieve the file, that client is
no longer suitable, e.g. because it no longer has the file.