store: Speed queries of type A up by adding a redundant clause
This helps greatly in speeding up one of the problematic queries we see in
production, bringing execution time down from 13s to 1.2s. Since the
additional query can slow down execution times dramatically when we are
looking for a large number of children, we only add the clause for small
numbers of children.

For now, that number is tuneable but should eventually become a constant