graph, graphql: Better error message when we decline to run a query
We used to always produce 'query too expensive' when the load manager
decided to decline running a query, which made users think there was
something wrong with their query.

We now return 'query too expensive' for queries that have been blocked
wholesale because we decided they were too expensive. For queries that we
will not run because the service is overloaded, we now return an error that
says that and makes it clear that that is a temporary condition.