all: Move building a QueryStore to a QueryStoreManager
We will have to have a separate store to look up block hashes in the
QueryStore. The only place that has access to both a block store and a
sdeployment store is the NetworkStore. We only want to export limited
functionality from the NetworkStore to the GraphQlRunner, hence a new trait
to get a QueryStore