*: Add Require a block number in `proofOfIndexing` queries
This is to avoid `null` POIs if the block hash isn't found in the block
store (which could be the case on one indexer but not on another). This
can lead to indexers losing rewards despite indexing a subgraph
correctly.

By requiring e.g. the indexer agent to always provide block number and
hash in `proofOfIndexing` queries, we can compute the POI on any
indexer, regardless of what is in their block store. This is easier than
resolving the block hash -> number on-demand in the `proofOfIndexing`
resolver. The agent already knows both values anyway.

Part of fixing https://github.com/graphprotocol/indexer/issues/186.