store: Dynamically load new ChainStores
Previously, the code assumed that all chains are known from the
configuration. That does not work if a chain was previously configured, but
is not currently configured (and leads to errors from the status/explorer
API). It also does not work when the configuration changes, but query nodes
are not restarted.

With this change, unknown chains are looked up in the database, so that,
for example, query nodes do not need to be restarted when the indexing
configuration changes.