store: Embed contract source in dynamic data source
For now, we write the source for a dds into two places redundantly. In
future changes, we will copy existing data into the new attributes and
start reading from there. Since the copying is very slow on the hosted
service, we split the migration into the new schema up into multiple steps.