all: Make the block number in EthereumBlockPointer a BlockNumber
We can't handle u64 block numbers - any interaction with the store with a
block number that does not fit into an i32 will blow up. Rather than do the
conversion of u64 to BlockNumber adhoc and in many places, force the code
to do those conversions when we take in a block number from an external
source.

The code could be tightened up further by using a newtype for BlockNumber,
but that is left for another change.