node, store: Do not block on eth adapters during startup
When an eth adapter would not return the net version or genesis block hash,
we would retry indefinitely. We now only try for 30s. If we can not get the
net version and genesis block hash during that time, we continue without
it. For networks we have used before, we assume that information has not
changed. For new networks, we fail startup.