all: Clarify that ingesting blocks always proceeds block-by-block
The code looked like we could try to find multiple missing ancestors for a
chain head candidate. But the logic of a chain dictates that we can only
ever find one block that we are missing when checking whether we have a
long enough chain in our database. Therefore, the vecs of missing parents
that we were passing around could only ever contain 0 or 1 element.

The code now clearly spells out that we will only ever know about one
missing parent at a time.