store: Avoid unnecessary updates in call_meta
Profiling data suggests that for busy chains with lots of subgraphs
indexing the same contracts, inserting new calls can take 300-400ms because
all the conflicting inserts need to wait for locks to do the update. With
this change, updates are performed at most once a day which should reduce
the amount of time indexing has to wait for these cache operations.