graph, node, store: Fix some inaccuracies with `graphman info`
The status that `graphman info -s` reported did not take into account that
we might have multiple copies of the same deployment in the store; in
particular, it reported the subgraph head pointer for the active deployment
for all copies.

This change fixes that, and also includes which copy is active in the output.