Added tracing logs when a giving deployment is assigned/not assigned to a given node
While doing some development, I had a misconfiguration where my `node_id` was not in sync with my configuration file.

This led to a deployment to be immediately removed on the receiving node leading to the deployment not starting.

To alleviate that problem a bit, I added some log lines that should make it clearer about what is happening and more especially why it's happening.

I choose to go with `debug` level but willing to switch to a `trace` level if you think it would be more appropriate.