core,chain,graph,node,store: added `FirehoseBlockStream` implementation of `BlockStream`
This PR adds a new `BlockStream` implementation called `FirehoseBlockStream` that is a built-in replacement for the existing `PollingBlockStream`. If your config has a provider of `type = firehose` (needs to use a new format for the provider definitions) **and** your started the `graph-node` with the environment variable `FIREHOSE_PRIORITY=true`, when a subgraph is syncing, a `FirehoseBlockStream` will be used to feed blocks to the subgraph deployed.

Left:
- [ ] More integration tests and behavior validation (dynamic data sources is one of them)
- [ ] Check how `FirehoseBlockStream` should be enabled in the future (by default if present, keep and env var, etc.)
- [ ] Review & addressing feedback