store: Allow setting min_idle and idle_timeout
This introduces two new environment variables that allow setting the
minimum number of idle connections and the timeout when a connection is
considered idle. If the variables are not set, the behavior defaults to the
current behavior of keeping all connections and never reaping idle ones,
and an idle timeout of 10 minutes.