store: Make DeploymentStore::transact_entity_operations take references
This is in preparation of retrying operations: if we pass owned data, we
need to clone before every attempt. Instead of forcing a clone in the
common case of the first try succeeding, work with references.

This also requires that we use Cow<Entity> because in rare cases we have to
modify the entity to add fulltext search fields; in the common case where
there is no fulltext search, we do not want to clone.