graph, firehose: fix panick because of re-use of an already completed stream
The state machine of the FirehoseBlockStream was not properly resetting the state to disconnected on error leading to the usage of an already completed future leading to a panic.

All code paths are not properly returning the state to Disconnected so the stream can be re-created to contact the remote Firehose endpoint.