eth: remove the traces requirement on triggers_adapter when Firehose provider will be used (#2847)
When Firehose is used to provide the data in the `graph-node`, the `traces` requirement is not necessary as the `traces` will be provided by `Firehose` itself.

This PR removes the `traces` requirement before retrieving the cheapest adapter.

Adjusted some error logging so that downstream messages are easier to read and improved Firehose connection error reporting.

It appears the `.context(...)` causes the error to not be shown in full which does not help debugging. Will need to see why later but for now, simply removed the `.context(...)` call.