blockchain, near: Remove near::MappingTrigger
The distinction between TriggerData and MappingTrigger
seems mostly accidental, this moves towards unification
by making them both NearTrigger for Near.

This is done by adding a TriggerWithHandler, which is
a reusable type that carries most MappingTrigger responsibilities.
This ended up also simplifying the ethereum MappingTrigger a bit,
which we eventually may also unify with the TriggerData.