node: Make `graphman rewind` safer to use
When rewinding a subgraph, it's important that the subgraph is not indexing
at the same time. The `rewind` command now automatically pauses subgraphs
before rewinding and resumes them afterwards.

It is now also possible to pass multiple subgraphs to `rewind`, and to
specify them by name, schema name, or deployment hash.