runtime: Improved performance of `to_asc_bytes` implementation of AscType derive macro
Currently to serialize a structure to a bytes buffer, we start with an empty buffer, extend it multiple time, one for each field of the struct and then checking that the byte count is equal to the memory size of struct in Rust memory. This creates N allocations, N being determined on the implementation of `extend_from_slice` but at worst could be `len(fields)`.

Since we know the size of bytes required in advanced, we should instead directly allocate the right amount of bytes. This will remove some useless allocations and should improve performances.