near: Fixes more alignment issue around NEAR Rust Asc bindings (#2913)
* near: Fixes more alignment issue around NEAR Rust Asc bindings

It appears some structure are unfixable, any order of the fields leads to extra padding in Rust memory.

AssemblyScript and `repr(C)` in Rust does not follow exactly the same rules always. One caveats is that some struct are packed in AssemblyScript but padded for alignment in `repr(C)` like a struct `{ one: AscPtr, two: AscPtr, three: AscPtr, four: u64 }`, it appears this struct is always padded in `repr(C)` by Rust whatever order is tried. However, it's possible to packed completely this struct in AssemblyScript and avoid any padding.

To overcome those cases where re-ordering never work, you will need to add an explicit _padding field to account for missing padding and pass this check.