graph, store: Filter subscriptions before sending them
We currently send store events and then filter them on the receiver
side. That means that if one receiver is slow, its buffer will fill up with
events, inhibiting progress, even though most of those buffered events will
be discarded when the receiver gets around to looking at them.

With this change, we discard events before sending them to reduce the
effect that slow receivers have on processing of events.