graph: fix FirehoseBlockStream re-connection when the cursor is "in memory"
So we've an architectural problem here where we cannot just save the response's cursor as the active cursor when a response is received in `FirehoseBlockStream`. It's actually the consumer of the stream that should inform us if a block was correctly fully consumed and that the latest cursor should be saved.

Imagine the following situation where we save the cursor right before sending it to the consumer in the `FirehoseBlockStream`. Now, the consumer of our stream receives it properly but fails without having fully processed the block (for example a bug bailed at the very first transaction of a block but there is still N transactions to process in it). Now if this stream reconnects with latest saved cursor, we will actually skip a block because the next block received will be right the saved cursor while we did not fully consumed the block the cursor was at.

To achieve that, we add a `notify_block_consumed` notification like method on the `BlockStream` trait with a default implementation that do nothing. When the instance manager consumed the block completely, it notifies the `BlockStream` implementation that the block was correctly processed. The `FirehoseBlockStream` when it's about to send a block to the consumer, saves the current cursor in its state. When the `notify_block_consumed` is received, it then switch the last seen cursor with the active one ensuring that we do not skip any blocks at all.