core: when block stream is canceled, we should return early on
It appears right now that subgraph keeps running when cancelation operation occurs (at least for FirehoseBockStream, not sure for PollingBlockStream).

Indeed, when a subgraph is canceled, we are currently break from the inner loop just to go back to it and re-try the block stream. However, when the deployment has been canceled, we should exit early on and simply quit at that location.