all: Change GraphQL output to follow the order of fields in the query
That is demanded by the GraphQL spec. We need to have a predictable order
to ensure attestations remain stable, though this change breaks
attestations since it changes the order in which fields appear in the
output from alphabetical (however a `BTreeMap` orders string keys) to the
order in which fields appear in a query.

It also allows us to replace `BTreeMaps`, which are fairly memory
intensive, with cheaper `Vec`.

The test changes all reflect the changed output behavior; they only reorder
fields in the expected output but do not otherwise alter the tests.

Fixes https://github.com/graphprotocol/graph-node/issues/2943