Avoid parsing X triggers if the filter is empty
When a filter is empty, it means it cannot match anything. By guarding the parsing of the triggers upfront, we avoid looping through lot's of element in the blocks, this is especially true for full blocks that contain all the transactions.

Remove a `clone()` that can be delayed to later at the same time.

Fixes #3080