chain/near: added dynamic data source support
#### WIP

This cannot be merged as-is. While doing the work, I realized that how data source are stored (via `StoredDynamicDataSource) and queried is still chain specific and tied to Ethereum.

Some thinking and discussions will be required to decide what will be the best approach to make them chain agnostic and how we are going to have this properly handle backward compatibility for existing stored dynamic data source.

Right now in this PR, I "hijack" the current format and I'm able to make NEAR dynamic data source works.