store: Fix the logic for determining if a graft was pending (again)
The previous version of the code was being too clever: the signal for
whether a graft is pending is whether the subgraph's block ptr is null or
not. The check whether the block ptr had progressed beyond the graft point
was too defensive and made it impossible to graft onto block 0, as well as
breaking a test.