chain/ethereum: Fix broken deployment_eth_rpc_* metrics registration
I noticed the following errors being logged frequently:

    ERRO registering metric [deployment_eth_rpc_request_duration] because it was already registered, component: MetricsRegistry
    ERRO registering metric [deployment_eth_rpc_errors] because it was already registered, component: MetricsRegistry

I also noticed that graph-node wouldn't correctly observe these metrics
because of this. The reason for the errors was that the metrics were
recreated each time a new block stream for a deployment was created.

This commit changes it so we remember the `SubgraphEthRpcMetric`
associated with each deployment instead of trying to recreate them each
time.