graph, graphql: Represent map keys in query results as Box<str>
We don't need the mutability that String allows for here, and a Box<str>
has an overhead of 16 bytes for storing a string vs the 24 bytes overhead
that a String requires.

This is particulalry important since query responses contain a lot of short
strings; e.g., to represent the key `id` using a String requires 26 bytes,
but with Box<str> it only requires 18 bytes.