graph, graphql, store: Use the correct parent type for type d queries
When we stitch the result of a query of type d back into the parents, we
need to make sure that the parent ids we pass through the query use the
type of the parent's id field. Previously, the code was using the type of
the child's id field but that has no relation to the correct type.

Fixes https://github.com/graphprotocol/graph-node/issues/3600