store: Make it possible to restart a failed writer
So far, if the writer for a subgraph encounterd an error, it would stick
around as 'poisoned' for as long as the process lived. That made it
impossible to restart a subgraph (possibly after clearing some external
error condition) With this change, `SubgraphStore.writable` will create a
new writer if the old one has been poisoned.