store: Be considerate of replicas when copying
Copying of subgraphs can create a lot of data to be replicated, which can
make replicas fall hopelessly behind. With this change, the copy code
checks whether there are replicas that are too far behind and backs off of
copying for a bit to give them a chance to catch up.