store: Go back to the less clever but correct way of pruning
The change in aea5b577 has a logic error in that we look at a fixed number
of rows from the table, filter out the ones we don't need to copy, and then
insert the remaining rows. The error is that the logic assumes that if we
do not insert any rows that we are done with copying, but that's wrong,
since not inserting rows now just means that no rows in the batch needed to
be copied even though a subsequent batch might still have such rows.

This reverts commit aea5b577db0fc5f08e8de17267c859186d917448.