The `SubstreamsBlockStream` was using `request.clone()` side stepping latest cursor value
The `request.clone()` does not correctly use the `latest_cursor` value which is the valid up to date in memory cursor to use on re-connection. This led to poisining error in `graph-node` where the same block was processed multiple time because the cursor was not correctly used.

Fixed by moving the request creation directly where it's needed which will use the correct up to date `latest_cursor` value now.