Added `call_only` feature on provider to be able to be picked up only for `eth_call` handling
I had to made some modification to how exactly `triggers_adapter` is actually called and implemented.

Indeed, in `instance_manager`, we retrieved a `triggers_adapter` based on the subgraph requested capabilities. This was causing errors when defining the provider with 'archive' as 'call_only'.

The problem is that at that point in `instance_manager`, nobody is actually going to use `archive` directly, only `traces` is of interset. But `instance_manager` is chain agnostic, so I couldn't implement the logic there. To keep the validation of the features of a subgraph, I introduced `is_capabilities_fulfilled` on `Chain` trait. That way, the `instance_manager` can validate the features.

And the `triggers_adapter` on Ethereum implementation was changed to remove the need for 'archive' capabilities because it's not needed there. Only when actually getting a provide to perform the `eth_call` of the subgraph that we need `archive`.

Happy to discuss those change over a call if needed.