graph, store: Restructure how pruning is done
We used to perform pruning in two stages, by first copying final entities
for all tables, then copying nonfinal entities and switching for all
tables.

It is better to do this loop the other way around: we now go
table-by-table, and for each of them do the nonfinal copy, then the final
copy. This makes an ongoing prune operation less visible, since the
subgraph writer can write in between the final copying for each table.