all: Make `graphman chain call-cache remove` safer
The command defaulted to the dangerous operation of clearing the entire
cache. Now, the entire cache can only be cleared by passing the annoyingly
named `--remove-entire-cache`.

In a similar vein, it used to be ok to omit --from or --to which would
default to the first or last block. They now need to both be specified to
make the user be clearer about how much of the call cache they are going to
remove.