store: Remove ability to recount entities (#4406)
We used to send very clever SQL to the database on every block to make it
possible to force a recount of the entities in a subgraph by setting the
entity_count to -1 or null. The SQL that got sent could be very
complicated, and might introduce some overhead.

This facility hasn't been used in a very long time, and since it might
cause an overhead, it's better to remove it. If recounting entities is
something that we do need, it should be added as a `graphman` command and
not be something that might cause headaches on the busiest code path.