store: Estimate amount of history based on actual amount of history
We used to determine the total number of blocks in a subgraph based on its
latest and earliest blocks. With ongoing pruning, the earliest block is
updated every time we prune, even though the logic in PruneRequest.strategy
might have us actually not do anything. That leads to a situation where we
think the subgraph contains much fewer blocks than it really does, and we
therefore underestimate how much of its data is historical.

We now remember for each table the block at which we actually pruned, which
might be long before the subgraph's earliest block, and use that to
determine how many blocks are present.

As an example, assume we want to keep 100 blocks of history, in a subgraph
that is at block 1000 and earliest block 800 and a table that was last
pruned at block 500. Previously, we would have estimated that 50% of the
table is historical, when in reality 80% is historical.