store: Treat 'no connection to the server' as db not available
So far we treated the Postgres error "no connection to the server" as not
indicating that the database was not available. It's a little murky what
exactly that error indicates, but it seems to indicate that an existing
connection got killed, e.g., because the database crashed.

We now treat this as an indication that the database is not available,
which will trigger the right action further up in the stack. For example,
on indexing, db operations will be retried instead of causing the subgraph
to fail. Queries will be aborted with an error message.