graph, store: Get rid of unneeded removal of __typename from entity
The code in EntityCache that removed `__typename` from entities is not
needed since the store does not set that attribute on the code paths that
the EntityCache uses.

This also adds a test for that. There are some tests that will fail
coincidentally if entities contain a `__typename`, but it's better to have
an explicit test for that.