perf: Avoid cloning cached entities
This should be a measurable performance improvement. And the lifetime
adjustments surprisingly didn't cause much trouble. The only case
where a clone is still required is if the entity needs to be mutated
by applying pending updates, so we use a `Cow` to handle that.
The entity sorting step is also faster now, because it no longer
requires cloning the entity keys, rather we just sort the interned
values.