all: Introduce subgraph-specific settings
It is now possible to set the history_blocks for a subgraph from a config
file that matches on subgraph names.

Settings are matched in the order in which they appear in the config file.

`graphman config check` checks that config file for validity if one is set
in the environment