delete 'shallow' fh eth blks on rpc+ingestor start
when launching grpah-node with an ingestor on an eth chain with an RPC
provider (i.e. not firehose), this will delete the 'shallow' blocks in
chainX.blocks, so they do not cause issues to the RPC provider.

It only runs on ethereum chains where firehose is not used and when
ingestor is active.