graph, graphql: Use correct parent id list when optimizing child query
This reverts commit 906f5e69881df1975115934a8715801d04f92e1e. That commit
caused a bug because of a confusion about which list of parent ids to
use. If we had entities without children followed by ones with children, we
would join the wrong children to parents.

We now use the correct list of parent ids, the one that already has entries
with no children filtered out.