all: Introduce an `Id` type to handle id values
We used to pass ids around mostly as strings, though that was never quite
clear from the code and it may have also been the case that in some
instances we passed ids around as bytes when the type in the GraphQL schema
was Bytes.

This PR makes it a lot clearer which type of id we are dealing with and
tries to restrict conversions from/to strings for Bytes ids to places where
we take them in from the outside, mostly mappings, substreams, and reading
and writing from/to the database.

It should make it possible to add more functionality and avoid conversion
bugs we encountered in the past as well as make it easier to support other
types of id, like integers.