store: Change how we pass in parents and children for type C queries
We used to pass the children as one big matrix which had the disadvantage
that we couldn't use bind variables and had to pass in literal values.

With this way of passing in the parent/children pairs, we need 2 bind
variables per parent, which restricts us to about 30k parents, which is a
pretty unreasonable number of parents.