perf(store): Avoid the BRIN index for queries that filter by id
A query has a where clause which filters by id is best served
by the GiST index on `(id, block_range)`. However when the table
is set to account-like, the BRIN index is sometimes chosen by
the query planner even though it can have much worse performance.

We've encountered this repeatedly on some large subgraphs and even had
to turn off `account-like` on some tables to resolve the situation,
which is of course not ideal because some queries do benefit from it.
This happened even with BRIN indexes using using `*_multi_ops`.

This proposes addressing this by not adding the BRIN-friendly
clauses on queries that we know are filtering by id. Which include type C and
D queries, and top-level queries explicitly filtering by `id`.

Impact to be tested in real-world scenarios.