feat(starknet): move away from `FieldElement` for data source
Currently, the native `FieldElement` type is used for Starknet data
sources, which uses the Montgomery representation internally for fast
finite field mulitplication, at the cost of slower (de)serialization.
However, in `graph-node`, this type is only used for equality
comparison, which calls for a type that's backed by a `u8` array
instead.

In fact, the `graph::blockchain::DataSource` trait requires the return
type of its `address()` method to be `Option<&[u8]>`, which necessitates
the use of `u8`-backed types (currently the Starknet data source simply
returnes `None`).

This commit adds a new type `Felt` that's a wrapper around `[u8; 32]`,
and replaces any `FieldElement` occurrence. The `starknet-ff` dependency
is also removed.