graph: Fix an issue with interfaces implementing interfaces
The problem was in `api_schema::id_type_as_scalar`: that tried to determine
the id type for an interface in completely the wrong way by looking at the
interfaces that interface implemented. It then tried to look that up as an
object type which would always fail.

We now take advantage of the precomputed `types_for_interfaces` map; most
of the changes are just to thread the `Schema` through to the right place
instead of just a `Document`