core, graph, store: Do not use recent blocks cache for chain head
With recent changes where the recent blocks cache gets populated on index
nodes that are not the block ingestor (#5059), the recent blocks cache can
have a notion of chain head pointer that is very far behind the actual
chain head pointer. In fact, there is no safe way to get the real chain
head pointer from the cache.

A side-effect of using the wrong chain head pointer is that a subgraph gets
marked as synced too early, which in turn turns off write batching.

Caching the chain head pointer is also a bit of a premature optimization as
looking up the chain head pointer should be very cheap as it does a lookup
in the very small `ethereum_networks` table.

If we do want to cache the chain head pointer, we will have to use the
ChainHeadUpdate notifications to update such a cache.