Fix data source rewind across more than one block (#5083)
* test: Add failing test for dds revert

* refactor: Remove 'Cancel the stream for real'

The `Action::Restart` will already drop the block stream and replace the guard in
the `instances` map, cancelling it in a very real way.

* fix(runner): Rationalize usage of `fn revert_state`

When reverting, the target block should be the target block of the revert,
so pass `revert_to_ptr` as the parameter and change the function to
`fn revert_state_to`. When restarting, it should revert to the last
known good block, but not actually revert it. The usage in the `Err`
case was redundant with the usage during restart.

* fix(revert): Fix long reverts of private dds