v0.34.0 Release (#5129)
* cargo: update workspace crates version to v0.34.0

* update NEWS.md

* NEWS.md: Add a note about resolving lingering rewind bugs

* NEWS.md update indexer hints syntax

* Update NEWS.md

Co-authored-by: Leonardo Yvens <leoyvens@gmail.com>

* Update NEWS.md

Co-authored-by: Leonardo Yvens <leoyvens@gmail.com>

* Update NEWS.md

Co-authored-by: Leonardo Yvens <leoyvens@gmail.com>

* server: fix graphiql issue when querying subgraph names with multiple path segments

* feat(starknet): rename block_handlers to block_handler

Currently, an array of block handlers is allowed, but only the first one
is actually used. This can easily cause confusion for end users.

While this is technically a breaking change, it should be fine as the
support for Starknet isn't official yet.

* feat(starknet): move away from `FieldElement` for data source

Currently, the native `FieldElement` type is used for Starknet data
sources, which uses the Montgomery representation internally for fast
finite field mulitplication, at the cost of slower (de)serialization.
However, in `graph-node`, this type is only used for equality
comparison, which calls for a type that's backed by a `u8` array
instead.

In fact, the `graph::blockchain::DataSource` trait requires the return
type of its `address()` method to be `Option<&[u8]>`, which necessitates
the use of `u8`-backed types (currently the Starknet data source simply
returnes `None`).

This commit adds a new type `Felt` that's a wrapper around `[u8; 32]`,
and replaces any `FieldElement` occurrence. The `starknet-ff` dependency
is also removed.

* feat(starknet): data source validation

* docs: document the `Felt` type

* refactor: remove `unsafe` in `starknet_keccak`

* update changelog link in NEWS.md

---------

Co-authored-by: David Lutterkort <lutter@watzmann.net>
Co-authored-by: Leonardo Yvens <leoyvens@gmail.com>
Co-authored-by: Jonathan LEI <me@xjonathan.dev>