all: Enable querying of aggregations
This required a major rework of the GraphQL query infrastructure as
prefetch now needs to translate from types in the API schema to types in
the InputSchema as the concrete types for the aggregations do not exist in
the API schema.

Before, we relied a lot on the fact that the API schema contained the same
types as the input schema with the exact same names and structure. But
conceptually, prefetch and query building deals with types from the input
schema as that describes how data is ultimately stored in the database.