store: Fix an issue in the aggregation rollup queries
We use 'order by timestamp' on the raw data to make sure that first and
last work correctly. But the name 'timestamp' was ambiguous as we select a
timestamp and the data table has one. The selected timestamp is the same
for all data points in the interval, whereas the timestamp from the data
table is the block's timestamp, which is what we need to sort by. This
change makes sure we sort by the timestamp from the data table as intended.