refactor: Remove hack in chain_head_listener
This was added in a time when we were having problems with
receiving chain head updates and were trying workarounds.
This one had a hypothesis of a deadlock in tokio Watch::send.
I was never able to corroborate this, no other tokio user ever
reported something like this and improvements have since been
made to the internals of `Watch`. So there isn't much justification
for keeping this hack.