store: Increase sleep time in write queue processing
We use sleep as a fallback in case the notification to write does not
work. Unfortunately, this leads to much smaller write batches than
configured for busy subgraphs as the check whether a batch should be
written seems to conflict with appending to a batch (we start a new batch
in that case)

By raising the safety sleep from 2s to GRAPH_STORE_WRITE_BATCH_DURATION we
make these collisions much less likely; there should not be any adverse
consequences. If the notifications don't work, we'd still wait at most the
configured time to write.