ipfs: Stop using the stat API
To reduce the total number of calls to IPFS, and
to generally simplify our interface on IPFS, this
stops using the `stat` API to check existence and
control file size. Both can be achieved efficiently
with the more usual cat api.

Relying on files/stat was questionable anyways,
since we don't know to what extent IPFS clients
actually enforce that the declared metadata size value
matches the real file size.