store: Change which indexes are created to support aggregations
For cumulative aggregations, the code was confused about the table on which
the index should be created. It needs to go on the aggregation table, not
the underlying source table. These indexes have a significant impact on the
performance of rollups, and having them on the wrong table makes rollups
very slow.

Fixes https://github.com/graphprotocol/graph-node/issues/5280