store: Check interface uniqueness in bulk
We used to run one query for each entity that was created or updated if it
was of a type that implemented an interface. These single-entity queries
can become quite time consuming during syncing.

With this change, we check all the entities of a given type with a single
query, so that uniqueness checking for interfaces causes as many queries as
there are types (with new/updated entities) rather than as many as there
are entities of such types.