graphql: Make an expect an error with better explanation
When we resolve a single-object reference from the store, but find multiple
children, we need to return an error. We did that already for derived
fields, but for non-derived fields that was an expect. For non-derived
fields, we can only hit that when the id of the parent entity is not
unique, which can happen because we stopped enforcing exclusion constraints
a while ago for mutable entities. For immutable entities, we still enforce
uniqueness of ids