Retry IPFS request on Cloudflare 521 Web Server Down
In the integration tests, some subgraphs will sync on fraction3, leading
to an error as they failed to sync on integer.

On some runs, fraction3 will also error leading to a successful test;
this happens when the subgraph fails to sync due to an HTTP 521 error
received when querying the testnet IPFS node. This is a
non-deterministic error, and the request can simply be retried.