Exclude full-text search columns from entity queries (#5693)
- Remove `tsvector` columns from `SELECT` clauses in entity queries
- Filter out full-text search columns from the list of selected columns

Full-text search columns (`tsvector` type) are used for indexing and
efficient text searching, but they are not part of the entity's data
model and are not meant to be directly queried or returned. This fixes a
bug where trying to load a tsvector column fails, as diesel does not
natively have tsvector support, and we don't explicitly handle the type
in `relational::value::OidValue::from_sql`.