store: Create postponed indexes non-concurrently for copy/graft
At the point where we create the postponed indexes during copying, nothing
else is writing to the subgraph and we can't be blocking a writer with a
normal 'create index'. Since concurrent index creation has to wait for all
previous transactions in the database to finish, the concurrent creation
can significantly slow down index creation and therefore how long the copy
takes.