graph, store: Add a timeout for each batch of data copying
When the data distribution in a table is bad (e.g., if it contains huge
arrays after a certain point) it can causes batches to run for many
hours. We now set a timeout and reset the batch size to 1 when that happens
so that we can slowly inch to a more reasonable batch size.