store: Try to spawn copy workers more frequently
We used to only try to spawn a new copy worker when we were finished
copying a table. But that can take a long time, and in between some
connections might have become available. We now check every few minutes if
a connection is available and spawn a new worker if it is