store: Change db setup strategy to guard better against races
The current database setup code was inerently racy when several nodes were
starting up as it relied on piecemeal locking of individual steps. This
change completely revamps the strategy we use: setup now takes a lock on
the primary, so that only one node at a time will run the setup code.