all: Filter out shards with pool size 0
Instead of dealing with disabled shards (shards that have a pool size of 0
configured), filter those shards out on startup and warn about them.

The end effect is that for that configuration, users will get an error of
'unkown shard' rather than 'shard disabled'. Since configuring a shard to
have no connections is kinda pathological, and leads to an error when it is
used either way, the code simplification is worth the slightly less helpful
error message.

Removing the 'disabled' state from pools has ripple effects to quite a few
other places, simplifying them a bit