store: Make sure we do not run setup twice for the same pool
With the previous code, we would run setup initially when creating all
pools, but they would not be marked as set up. On the first access to the
pool we would try to run setup again, which is not needed. This change
makes it so that we remember that we ran setup successfully when pools are
created