store: Do not use a fdw connection to check active_copies
Copying checks the active_copies table through the
primary_public.active_copies foreign table to determine whether the copy
has been cancelled and it should stop.

With a large number of copies running, that causes a large number of
postgres_fdw connections into the primary, which can overwhelm the primary.

Instead, we now pass the connection pool for the primary into the copy code
so that it can do this check without involving postgres_fdw.